--- chatroom/src/Chat.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-function Chat() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-
-export default Chat
