+++ chatroom/src/carousel/carousePlatforms.js
