+++ chatroom/src/carousel/carouselHardware.js
