+++ chatroom/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './css/index.css';
 import App from './App';
 import {BrowserRouter} from "react-router-dom";
 // import 'bootstrap/dist/css/bootstrap.min.css';
